# docs/api/api.yaml
...
openapi: 3.0.3
info:
  title: Shelfinity Library Management System API
  description: >
    RESTful API for managing library books, users, borrow/return requests,
    reservations, overdue tracking, email configuration, and reports.
    Authentication is delegated entirely to Keycloak (OIDC) — this API validates
    bearer tokens, it does not issue them or store credentials.
  version: 1.0.0
  contact:
    name: Shadow-Codex
servers:
  - url: http://localhost:9080/api
    description: Local development server
...
paths:
...
  /queues:
    post:
      tags: [Queue]
      summary: Submit a borrow, return, or registration-approval request
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateQueueItemRequest' }
      responses:
        '201':
          description: Queue item created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueueItemResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Queue]
      summary: List all queue items (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/QueueStatus' }
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/QueueType' }
      responses:
        '200':
          description: Queue items retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/QueueItemResponse' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /queues/my:
    get:
      tags: [Queue]
      summary: List the current user's queue items
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Queue items retrieved
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/QueueItemResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /queues/{id}:
    parameters:
      - $ref: '#/components/parameters/UuidId'
    get:
      tags: [Queue]
      summary: Get a queue item by ID
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Queue item retrieved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueueItemResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Queue]
      summary: Delete a queue item (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Queue item deleted }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /queues/{id}/status:
    parameters:
      - $ref: '#/components/parameters/UuidId'
    patch:
      tags: [Queue]
      summary: Approve or reject a queue item (admin only)
      description: >
        Behavior per SPEC.md §6.1/§6.2/§6.3, all resolved: on BOOK_BORROW
        approval, decrements book availability (409 if none left) and sets a due
        date; on BOOK_RETURN approval, increments availability and promotes the
        oldest active reservation if one exists, resetting its claim-window
        expiry; on USER_REGISTRATION approval, activates the user account.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateQueueItemRequest' }
      responses:
        '200':
          description: Queue item updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QueueItemResponse' }
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------- Reservations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued OIDC access token.
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated but not permitted (e.g. admin-only endpoint)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
...
    QueueType:
      type: string
      enum: [USER_REGISTRATION, BOOK_BORROW, BOOK_RETURN, BOOK_RESERVATION]
      description: >
        BOOK_RESERVATION exists in the enum but is legacy/unused — reservations
        are modeled as their own resource. See SPEC.md §5.3.
    QueueStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    QueueItemResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/QueueType' }
        status: { $ref: '#/components/schemas/QueueStatus' }
        userKeycloakId: { type: string }
        bookId: { type: string, format: uuid, nullable: true }
        description: { type: string }
        adminRemark: { type: string, nullable: true }
        dueDate: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time, nullable: true }
        processedAt: { type: string, format: date-time, nullable: true }
        processedBy: { type: string, nullable: true }
    CreateQueueItemRequest:
      type: object
      required: [type, description]
      properties:
        type: { $ref: '#/components/schemas/QueueType' }
        bookId:
          type: string
          format: uuid
          description: Required for BOOK_BORROW / BOOK_RETURN
        description: { type: string }
    UpdateQueueItemRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/QueueStatus' }
        adminRemark: { type: string }
...